+++ pkg/infinity/client.go
@@ -1,6 +1,7 @@
 package infinity
 
 import (
+	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/base64"
@@ -13,6 +14,10 @@ import (
 	"net/url"
 	"strings"
 	"time"
+
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/clientcredentials"
+	"golang.org/x/oauth2/jwt"
 )
 
 type Client struct {
@@ -58,6 +63,46 @@ func NewClient(settings InfinitySettings) (client *Client, err error) {
 	httpClient := &http.Client{
 		Transport: transport,
 		Timeout:   time.Second * time.Duration(settings.TimeoutInSeconds),
+	}
+	if settings.AuthenticationMethod == "oauth2" {
+		if settings.OAuth2Settings.OAuth2Type == "client_credentials" {
+			oauthConfig := clientcredentials.Config{
+				ClientID:       settings.OAuth2Settings.ClientID,
+				ClientSecret:   settings.OAuth2Settings.ClientSecret,
+				TokenURL:       settings.OAuth2Settings.TokenURL,
+				Scopes:         []string{},
+				EndpointParams: url.Values{},
+			}
+			for _, scope := range settings.OAuth2Settings.Scopes {
+				if scope != "" {
+					oauthConfig.Scopes = append(oauthConfig.Scopes, scope)
+				}
+			}
+			for k, v := range settings.OAuth2Settings.EndpointParams {
+				if k != "" && v != "" {
+					oauthConfig.EndpointParams.Set(k, v)
+				}
+			}
+			ctx := context.WithValue(context.Background(), oauth2.HTTPClient, httpClient)
+			httpClient = oauthConfig.Client(ctx)
+		}
+		if settings.OAuth2Settings.OAuth2Type == "jwt" {
+			jwtConfig := jwt.Config{
+				Email:        settings.OAuth2Settings.Email,
+				TokenURL:     settings.OAuth2Settings.TokenURL,
+				PrivateKey:   []byte(strings.ReplaceAll(settings.OAuth2Settings.PrivateKey, "\\n", "\n")),
+				PrivateKeyID: settings.OAuth2Settings.PrivateKeyID,
+				Subject:      settings.OAuth2Settings.Subject,
+				Scopes:       []string{},
+			}
+			for _, scope := range settings.OAuth2Settings.Scopes {
+				if scope != "" {
+					jwtConfig.Scopes = append(jwtConfig.Scopes, scope)
+				}
+			}
+			ctx := context.WithValue(context.Background(), oauth2.HTTPClient, httpClient)
+			httpClient = jwtConfig.Client(ctx)
+		}
 	}
 	return &Client{
 		Settings:   settings,