+++ src/types.ts
@@ -6,7 +6,20 @@ export interface GlobalInfinityQuery {
   id: string;
   query: InfinityQuery;
 }
+export type AuthType = 'none' | 'basicAuth' | 'oauthPassThru' | 'oauth2';
+export type OAuth2Type = 'client_credentials' | 'jwt' | 'others';
+export type OAuth2Props = {
+  oauth2_type?: OAuth2Type;
+  client_id?: string;
+  email?: string;
+  private_key_id?: string;
+  subject?: string;
+  token_url?: string;
+  scopes?: string[];
+};
 export interface InfinityOptions extends DataSourceJsonData {
+  auth_method?: AuthType;
+  oauth2?: OAuth2Props;
   tlsSkipVerify?: boolean;
   tlsAuth?: boolean;
   serverName?: string;
@@ -21,6 +34,8 @@ export interface InfinitySecureOptions {
   tlsCACert?: string;
   tlsClientCert?: string;
   tlsClientKey?: string;
+  oauth2ClientSecret?: string;
+  oauth2JWTPrivateKey?: string;
 }
 export interface SecureField {
   id: string;